requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
flask-jwt-extended==4.5.1
flask-marshmallow==0.15.0
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
PyMySQL==1.1.0
Werkzeug==2.3.7
passlib==1.7.4
